refactor(javm-cap): move CapRef into cache.rs and make new() private
`CapRef` is a cache-internal lifetime handle; its only legitimate
production constructor is `CacheDirectory::put_instance`. Hosting it
in `cap.rs` forced the constructor to be `pub(crate)` so cross-module
access could work. Moving the type into `cache.rs` lets the
constructor be plain module-private (`fn new`) — strictly tighter
than `pub(crate)`.

```rust
// cache.rs
pub struct CapRef { id: u64, rc: Arc<()> }
impl CapRef {
    fn new(id: u64) -> Self { ... }  // module-private
    pub fn id(&self) -> u64 { ... }
    pub fn strong_count(&self) -> usize { ... }
}
```

`cap.rs` now `use super::cache::CapRef;` for the `CapHashOrRef::Ref`
variant. The module-doc block about CapRef-as-handle moves to the
new type definition. `lib.rs` re-exports `CapRef` from `cache`
instead of `cap` (the canonical workspace path `javm_cap::CapRef`
is unchanged). The one external module that referenced
`javm_cap::cap::CapRef` (`nub-arch-x86/state_cache.rs`) is updated
to `javm_cap::cache::CapRef`.

Verified: workspace build/clippy/fmt clean; `cargo test -p javm-cap`
green; smoke green on both backends.
